+++ lib/src/messages/file_message.dart
@@ -18,12 +18,14 @@ class FileMessage extends Message {
     required User author,
     int? createdAt,
     required String id,
+    this.isLoading,
     Map<String, dynamic>? metadata,
     this.mimeType,
     required this.name,
     String? remoteId,
     Message? repliedMessage,
     String? roomId,
+    bool? showStatus,
     required this.size,
     StatusType? status,
     MessageType? type,
@@ -39,6 +41,7 @@ class FileMessage extends Message {
           remoteId,
           repliedMessage,
           roomId,
+          showStatus,
           status,
           type ?? MessageType.file,
           updatedAt,
@@ -51,11 +54,16 @@ class FileMessage extends Message {
     required User author,
     int? createdAt,
     required String id,
+    this.isLoading,
     required PartialFile partialFile,
     String? remoteId,
     Message? repliedMessage,
     String? roomId,
+
     StatusType? status,
+
+    bool? showStatus,
+    
     int? updatedAt,
     MessageDeleteType? deleteType,
     int? deletedAt,
@@ -71,6 +79,7 @@ class FileMessage extends Message {
           remoteId,
           repliedMessage,
           roomId,
+          showStatus,
           status,
           MessageType.file,
           updatedAt,
@@ -91,15 +100,22 @@ class FileMessage extends Message {
   /// both metadatas will be merged into one Map, where keys from a passed
   /// metadata will overwite keys from the previous one.
   /// [previewData] is ignored for this message type.
-  /// [remoteId] and [updatedAt] with null values will nullify existing value.
-  /// [status] and [uri] with null values will be overwritten by previous values.
+  /// [isLoading], [remoteId], [showStatus] and [updatedAt] with null values will nullify existing value.
+  /// [author], [createdAt], [status] and [uri] with null values will be overwritten by previous values.
   /// [text] is ignored for this message type.
   @override
   Message copyWith({
+    User? author,
+    int? createdAt,
+    bool? isLoading,
     Map<String, dynamic>? metadata,
     PreviewData? previewData,
     String? remoteId,
+
     StatusType? status,
+
+    bool? showStatus,
+    
     String? text,
     int? updatedAt,
     String? uri,
@@ -107,9 +123,10 @@ class FileMessage extends Message {
     int? deletedAt,
   }) {
     return FileMessage(
-      author: author,
-      createdAt: createdAt,
+      author: author ?? this.author,
+      createdAt: createdAt ?? this.createdAt,
       id: id,
+      isLoading: isLoading,
       metadata: metadata == null
           ? null
           : {
@@ -121,6 +138,7 @@ class FileMessage extends Message {
       remoteId: remoteId,
       repliedMessage: repliedMessage,
       roomId: roomId,
+      showStatus: showStatus,
       size: size,
       status: status ?? this.status,
       updatedAt: updatedAt,
@@ -136,6 +154,7 @@ class FileMessage extends Message {
         author,
         createdAt,
         id,
+        isLoading,
         metadata,
         mimeType,
         name,
@@ -149,6 +168,9 @@ class FileMessage extends Message {
         deleteType,
         deletedAt,
       ];
+
+  /// Specify whether the message content is currently being loaded
+  final bool? isLoading;
 
   /// Media type
   final String? mimeType;