+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 3.3.4
+
+- Add `isLoading` to the file message. Thanks @felixgabler for the PR!
+
+## 3.3.3
+
+- Add `author` and `createdAt` to the `copyWith` method. Thanks @felixgabler for the PR!
+
+## 3.3.2
+
+- Add `showStatus` to all messages. Thanks @arsamme for the PR!
+- Update to Flutter 2.10.4
+
 ## 3.3.1
 
 - Update to Flutter 2.10.2